Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai>=0.8.3
pillow>=10.2.0
pyautogui>=0.9.54
python-Levenshtein>=0.23.0
websockets>=12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
playwright>=1.40.0